+++ tests/test_delete_temporary_files.py
