+++ tests/test_generate_output_filename.py
