+++ tests/test_merge_bin_files.py
